TensorFlow: 1 step towards removing unnecessary colocate_with statements.
Simple placer no longer fails if the assign is placed on another
device than the variable; it logs INFO and says it ignores the placement.

This allows us to get rid of the colocation statements that are
superfluous in library code in python.  There are many more than
just this one, but these two are foundational ones that, if there
is an error, suggests that we still have some other work to do.  Ideally
this is a no-op.
Change: 120873104